Changed API to allow more incremental processing, and various refactoring.
include/extract.h:
    We now have functions for page begin/end and for adding spans and images to
    document, so we don't require intermediate data. However we can still read
    intermediate data if required with extract_read_intermediate().

    The API now makes use of abstract extract_t*, and allows joining of spans
    on a per-page basis, to reduce memory usage.

Makefile:
    Use plain grep instead of git grep when checking code for ssize_t etc, so
    that testing works on rsync-ed non-git checkouts.

    Better settings when running with memento.

    Use 'test' as main target, not 'tests'.

    Use $(mutool_e) for mutool's built-in extract functionality, so we can
    exclude some tests if not available.

src/docx_template_build.py:
    Force use of utf-8 when looking at template .docx files; fixes things on
    platforms with weird default encodings.

src/alloc.c:
    fixed bug in extract_malloc() - set *pptr to NULL on failure. This fixes
    memento squeeze failures.

Refactoring:
    Moved definitions of span_t, line_t etc from src/extract.c into
    src/document.h so that all code can use them directly if required -
    avoids previous awkward internal api boundary that forced callbacks into
    src/extract.c.

    Moved joining code into new src/join.c.

Other:
    Added extract_buffer_expanding* - provides an in-memory auto-growing
    contiguous write buffer.

    Removed various bits of disabled code.

    Change extract_write_content() to take an extract_buffer_t rather than a
    FILE*.

    Fixed bug in extract_buffer_write_internal() to make it return error if
    buffer is not a write buffer. Previously used to SEGV. Added a test for
    this.

    Removed some unnecessary includes.

    Various fixes for compilers that enforce C89 restrictions.

    Make extract_zip_close() take extract_zip_t** pzip.

+++ src/astring.h
@@ -7,7 +7,7 @@
 typedef struct
 {
     char*   chars;      /* NULL or zero-terminated. */
-    int     chars_num;  /* Length of string pointed to by .chars. */
+    size_t  chars_num;  /* Length of string pointed to by .chars. */
 } extract_astring_t;
 
 void extract_astring_init(extract_astring_t* string);
